Update ICS protos to ea545b490b53b8e6f422492b4fb2820ab0bc140b
