+++ src/prost/interchain_security.ccv.provider.v1.rs
@@ -16,24 +16,8 @@ pub struct MsgAssignConsumerKey {
 #[allow(clippy::derive_partial_eq_without_eq)]
 #[derive(Clone, PartialEq, ::prost::Message)]
 pub struct MsgAssignConsumerKeyResponse {}
-/// MsgRegisterConsumerRewardDenom allows an account to register
-/// a consumer reward denom, i.e., add it to the list of denoms
-/// accepted by the provider as rewards.
-#[allow(clippy::derive_partial_eq_without_eq)]
-#[derive(Clone, PartialEq, ::prost::Message)]
-pub struct MsgRegisterConsumerRewardDenom {
-    #[prost(string, tag = "1")]
-    pub denom: ::prost::alloc::string::String,
-    #[prost(string, tag = "2")]
-    pub depositor: ::prost::alloc::string::String,
-}
-/// MsgRegisterConsumerRewardDenomResponse defines the Msg/RegisterConsumerRewardDenom response type.
-#[allow(clippy::derive_partial_eq_without_eq)]
-#[derive(Clone, PartialEq, ::prost::Message)]
-pub struct MsgRegisterConsumerRewardDenomResponse {}
-/// MsgSubmitConsumerMisbehaviour defines a message that reports a misbehaviour
-/// observed on a consumer chain
-/// Note that the misbheaviour' headers must contain the same trusted states
+/// MsgSubmitConsumerMisbehaviour defines a message that reports a light client attack,
+///   also known as a misbehaviour, observed on a consumer chain
 #[allow(clippy::derive_partial_eq_without_eq)]
 #[derive(Clone, PartialEq, ::prost::Message)]
 pub struct MsgSubmitConsumerMisbehaviour {
@@ -49,8 +33,8 @@ pub struct MsgSubmitConsumerMisbehaviour {
 #[allow(clippy::derive_partial_eq_without_eq)]
 #[derive(Clone, PartialEq, ::prost::Message)]
 pub struct MsgSubmitConsumerMisbehaviourResponse {}
-/// MsgSubmitConsumerDoubleVoting defines a message that reports an equivocation
-/// observed on a consumer chain
+/// MsgSubmitConsumerDoubleVoting defines a message that reports
+/// a double signing infraction observed on a consumer chain
 #[allow(clippy::derive_partial_eq_without_eq)]
 #[derive(Clone, PartialEq, ::prost::Message)]
 pub struct MsgSubmitConsumerDoubleVoting {
@@ -188,36 +172,6 @@ pub mod msg_client {
                 );
             self.inner.unary(req, path, codec).await
         }
-        pub async fn register_consumer_reward_denom(
-            &mut self,
-            request: impl tonic::IntoRequest<super::MsgRegisterConsumerRewardDenom>,
-        ) -> std::result::Result<
-            tonic::Response<super::MsgRegisterConsumerRewardDenomResponse>,
-            tonic::Status,
-        > {
-            self.inner
-                .ready()
-                .await
-                .map_err(|e| {
-                    tonic::Status::new(
-                        tonic::Code::Unknown,
-                        format!("Service was not ready: {}", e.into()),
-                    )
-                })?;
-            let codec = tonic::codec::ProstCodec::default();
-            let path = http::uri::PathAndQuery::from_static(
-                "/interchain_security.ccv.provider.v1.Msg/RegisterConsumerRewardDenom",
-            );
-            let mut req = request.into_request();
-            req.extensions_mut()
-                .insert(
-                    GrpcMethod::new(
-                        "interchain_security.ccv.provider.v1.Msg",
-                        "RegisterConsumerRewardDenom",
-                    ),
-                );
-            self.inner.unary(req, path, codec).await
-        }
         pub async fn submit_consumer_misbehaviour(
             &mut self,
             request: impl tonic::IntoRequest<super::MsgSubmitConsumerMisbehaviour>,
@@ -293,13 +247,6 @@ pub mod msg_server {
             request: tonic::Request<super::MsgAssignConsumerKey>,
         ) -> std::result::Result<
             tonic::Response<super::MsgAssignConsumerKeyResponse>,
-            tonic::Status,
-        >;
-        async fn register_consumer_reward_denom(
-            &self,
-            request: tonic::Request<super::MsgRegisterConsumerRewardDenom>,
-        ) -> std::result::Result<
-            tonic::Response<super::MsgRegisterConsumerRewardDenomResponse>,
             tonic::Status,
         >;
         async fn submit_consumer_misbehaviour(
@@ -426,54 +373,6 @@ pub mod msg_server {
                     let fut = async move {
                         let inner = inner.0;
                         let method = AssignConsumerKeySvc(inner);
-                        let codec = tonic::codec::ProstCodec::default();
-                        let mut grpc = tonic::server::Grpc::new(codec)
-                            .apply_compression_config(
-                                accept_compression_encodings,
-                                send_compression_encodings,
-                            )
-                            .apply_max_message_size_config(
-                                max_decoding_message_size,
-                                max_encoding_message_size,
-                            );
-                        let res = grpc.unary(method, req).await;
-                        Ok(res)
-                    };
-                    Box::pin(fut)
-                }
-                "/interchain_security.ccv.provider.v1.Msg/RegisterConsumerRewardDenom" => {
-                    #[allow(non_camel_case_types)]
-                    struct RegisterConsumerRewardDenomSvc<T: Msg>(pub Arc<T>);
-                    impl<
-                        T: Msg,
-                    > tonic::server::UnaryService<super::MsgRegisterConsumerRewardDenom>
-                    for RegisterConsumerRewardDenomSvc<T> {
-                        type Response = super::MsgRegisterConsumerRewardDenomResponse;
-                        type Future = BoxFuture<
-                            tonic::Response<Self::Response>,
-                            tonic::Status,
-                        >;
-                        fn call(
-                            &mut self,
-                            request: tonic::Request<
-                                super::MsgRegisterConsumerRewardDenom,
-                            >,
-                        ) -> Self::Future {
-                            let inner = Arc::clone(&self.0);
-                            let fut = async move {
-                                (*inner).register_consumer_reward_denom(request).await
-                            };
-                            Box::pin(fut)
-                        }
-                    }
-                    let accept_compression_encodings = self.accept_compression_encodings;
-                    let send_compression_encodings = self.send_compression_encodings;
-                    let max_decoding_message_size = self.max_decoding_message_size;
-                    let max_encoding_message_size = self.max_encoding_message_size;
-                    let inner = self.inner.clone();
-                    let fut = async move {
-                        let inner = inner.0;
-                        let method = RegisterConsumerRewardDenomSvc(inner);
                         let codec = tonic::codec::ProstCodec::default();
                         let mut grpc = tonic::server::Grpc::new(codec)
                             .apply_compression_config(
@@ -622,9 +521,11 @@ pub mod msg_server {
         const NAME: &'static str = "interchain_security.ccv.provider.v1.Msg";
     }
 }
-/// ConsumerAdditionProposal is a governance proposal on the provider chain to spawn a new consumer chain.
-/// If it passes, then all validators on the provider chain are expected to validate the consumer chain at spawn time
-/// or get slashed. It is recommended that spawn time occurs after the proposal end time.
+/// ConsumerAdditionProposal is a governance proposal on the provider chain to
+/// spawn a new consumer chain. If it passes, then all validators on the provider
+/// chain are expected to validate the consumer chain at spawn time or get
+/// slashed. It is recommended that spawn time occurs after the proposal end
+/// time.
 #[allow(clippy::derive_partial_eq_without_eq)]
 #[derive(Clone, PartialEq, ::prost::Message)]
 pub struct ConsumerAdditionProposal {
@@ -634,26 +535,30 @@ pub struct ConsumerAdditionProposal {
     /// the description of the proposal
     #[prost(string, tag = "2")]
     pub description: ::prost::alloc::string::String,
-    /// the proposed chain-id of the new consumer chain, must be different from all other consumer chain ids of the executing
-    /// provider chain.
+    /// the proposed chain-id of the new consumer chain, must be different from all
+    /// other consumer chain ids of the executing provider chain.
     #[prost(string, tag = "3")]
     pub chain_id: ::prost::alloc::string::String,
     /// the proposed initial height of new consumer chain.
-    /// For a completely new chain, this will be {0,1}. However, it may be different if this is a chain that is converting to a consumer chain.
+    /// For a completely new chain, this will be {0,1}. However, it may be
+    /// different if this is a chain that is converting to a consumer chain.
     #[prost(message, optional, tag = "4")]
     pub initial_height: ::core::option::Option<
         super::super::super::super::ibc::core::client::v1::Height,
     >,
-    /// The hash of the consumer chain genesis state without the consumer CCV module genesis params.
-    /// It is used for off-chain confirmation of genesis.json validity by validators and other parties.
+    /// The hash of the consumer chain genesis state without the consumer CCV
+    /// module genesis params. It is used for off-chain confirmation of
+    /// genesis.json validity by validators and other parties.
     #[prost(bytes = "vec", tag = "5")]
     pub genesis_hash: ::prost::alloc::vec::Vec<u8>,
-    /// The hash of the consumer chain binary that should be run by validators on chain initialization.
-    /// It is used for off-chain confirmation of binary validity by validators and other parties.
+    /// The hash of the consumer chain binary that should be run by validators on
+    /// chain initialization. It is used for off-chain confirmation of binary
+    /// validity by validators and other parties.
     #[prost(bytes = "vec", tag = "6")]
     pub binary_hash: ::prost::alloc::vec::Vec<u8>,
-    /// spawn time is the time on the provider chain at which the consumer chain genesis is finalized and all validators
-    /// will be responsible for starting their consumer chain validator node.
+    /// spawn time is the time on the provider chain at which the consumer chain
+    /// genesis is finalized and all validators will be responsible for starting
+    /// their consumer chain validator node.
     #[prost(message, optional, tag = "7")]
     pub spawn_time: ::core::option::Option<
         super::super::super::super::google::protobuf::Timestamp,
@@ -679,8 +584,10 @@ pub struct ConsumerAdditionProposal {
     /// decimal number. For example "0.75" would represent 75%.
     #[prost(string, tag = "11")]
     pub consumer_redistribution_fraction: ::prost::alloc::string::String,
-    /// BlocksPerDistributionTransmission is the number of blocks between ibc-token-transfers from the consumer chain to the provider chain.
-    /// On sending transmission event, `consumer_redistribution_fraction` of the accumulated tokens are sent to the consumer redistribution address.
+    /// BlocksPerDistributionTransmission is the number of blocks between
+    /// ibc-token-transfers from the consumer chain to the provider chain. On
+    /// sending transmission event, `consumer_redistribution_fraction` of the
+    /// accumulated tokens are sent to the consumer redistribution address.
     #[prost(int64, tag = "12")]
     pub blocks_per_distribution_transmission: i64,
     /// The number of historical info entries to persist in store.
@@ -691,15 +598,16 @@ pub struct ConsumerAdditionProposal {
     /// The ID of a token transfer channel used for the Reward Distribution
     /// sub-protocol. If DistributionTransmissionChannel == "", a new transfer
     /// channel is created on top of the same connection as the CCV channel.
-    /// Note that transfer_channel_id is the ID of the channel end on the consumer chain.
-    /// it is most relevant for chains performing a sovereign to consumer changeover
-    /// in order to maintan the existing ibc transfer channel
+    /// Note that transfer_channel_id is the ID of the channel end on the consumer
+    /// chain. it is most relevant for chains performing a sovereign to consumer
+    /// changeover in order to maintan the existing ibc transfer channel
     #[prost(string, tag = "14")]
     pub distribution_transmission_channel: ::prost::alloc::string::String,
 }
-/// ConsumerRemovalProposal is a governance proposal on the provider chain to remove (and stop) a consumer chain.
-/// If it passes, all the consumer chain's state is removed from the provider chain. The outstanding unbonding
-/// operation funds are released.
+/// ConsumerRemovalProposal is a governance proposal on the provider chain to
+/// remove (and stop) a consumer chain. If it passes, all the consumer chain's
+/// state is removed from the provider chain. The outstanding unbonding operation
+/// funds are released.
 #[allow(clippy::derive_partial_eq_without_eq)]
 #[derive(Clone, PartialEq, ::prost::Message)]
 pub struct ConsumerRemovalProposal {
@@ -712,29 +620,34 @@ pub struct ConsumerRemovalProposal {
     /// the chain-id of the consumer chain to be stopped
     #[prost(string, tag = "3")]
     pub chain_id: ::prost::alloc::string::String,
-    /// the time on the provider chain at which all validators are responsible to stop their consumer chain validator node
+    /// the time on the provider chain at which all validators are responsible to
+    /// stop their consumer chain validator node
     #[prost(message, optional, tag = "4")]
     pub stop_time: ::core::option::Option<
         super::super::super::super::google::protobuf::Timestamp,
     >,
 }
+/// ChangeRewardDenomsProposal is a governance proposal on the provider chain to
+/// mutate the set of denoms accepted by the provider as rewards.
 #[allow(clippy::derive_partial_eq_without_eq)]
 #[derive(Clone, PartialEq, ::prost::Message)]
-pub struct EquivocationProposal {
+pub struct ChangeRewardDenomsProposal {
     /// the title of the proposal
     #[prost(string, tag = "1")]
     pub title: ::prost::alloc::string::String,
     /// the description of the proposal
     #[prost(string, tag = "2")]
     pub description: ::prost::alloc::string::String,
-    /// the list of equivocations that will be processed
-    #[prost(message, repeated, tag = "3")]
-    pub equivocations: ::prost::alloc::vec::Vec<
-        super::super::super::super::cosmos::evidence::v1beta1::Equivocation,
-    >,
+    /// the list of consumer reward denoms to add
+    #[prost(string, repeated, tag = "3")]
+    pub denoms_to_add: ::prost::alloc::vec::Vec<::prost::alloc::string::String>,
+    /// the list of consumer reward denoms to remove
+    #[prost(string, repeated, tag = "4")]
+    pub denoms_to_remove: ::prost::alloc::vec::Vec<::prost::alloc::string::String>,
 }
-/// A persisted queue entry indicating that a slash packet data instance needs to be handled.
-/// This type belongs in the "global" queue, to coordinate slash packet handling times between consumers.
+/// A persisted queue entry indicating that a slash packet data instance needs to
+/// be handled. This type belongs in the "global" queue, to coordinate slash
+/// packet handling times between consumers.
 #[allow(clippy::derive_partial_eq_without_eq)]
 #[derive(Clone, PartialEq, ::prost::Message)]
 pub struct GlobalSlashEntry {
@@ -754,7 +667,8 @@ pub struct GlobalSlashEntry {
     /// The provider's consensus address of the validator being slashed.
     /// This field is used to obtain validator power in HandleThrottleQueues.
     ///
-    /// This field is not used in the store key, but is persisted in value bytes, see QueueGlobalSlashEntry.
+    /// This field is not used in the store key, but is persisted in value bytes,
+    /// see QueueGlobalSlashEntry.
     #[prost(bytes = "vec", tag = "4")]
     pub provider_val_cons_addr: ::prost::alloc::vec::Vec<u8>,
 }
@@ -766,7 +680,8 @@ pub struct Params {
     pub template_client: ::core::option::Option<
         super::super::super::super::ibc::lightclients::tendermint::v1::ClientState,
     >,
-    /// TrustingPeriodFraction is used to compute the consumer and provider IBC client's TrustingPeriod from the chain defined UnbondingPeriod
+    /// TrustingPeriodFraction is used to compute the consumer and provider IBC
+    /// client's TrustingPeriod from the chain defined UnbondingPeriod
     #[prost(string, tag = "2")]
     pub trusting_period_fraction: ::prost::alloc::string::String,
     /// Sent IBC packets will timeout after this duration
@@ -774,7 +689,8 @@ pub struct Params {
     pub ccv_timeout_period: ::core::option::Option<
         super::super::super::super::google::protobuf::Duration,
     >,
-    /// The channel initialization (IBC channel opening handshake) will timeout after this duration
+    /// The channel initialization (IBC channel opening handshake) will timeout
+    /// after this duration
     #[prost(message, optional, tag = "4")]
     pub init_timeout_period: ::core::option::Option<
         super::super::super::super::google::protobuf::Duration,
@@ -792,8 +708,9 @@ pub struct Params {
     pub slash_meter_replenish_period: ::core::option::Option<
         super::super::super::super::google::protobuf::Duration,
     >,
-    /// The fraction of total voting power that is replenished to the slash meter every replenish period.
-    /// This param also serves as a maximum fraction of total voting power that the slash meter can hold.
+    /// The fraction of total voting power that is replenished to the slash meter
+    /// every replenish period. This param also serves as a maximum fraction of
+    /// total voting power that the slash meter can hold.
     #[prost(string, tag = "7")]
     pub slash_meter_replenish_fraction: ::prost::alloc::string::String,
     /// The maximum amount of throttled slash or vsc matured packets
@@ -806,23 +723,16 @@ pub struct Params {
         super::super::super::super::cosmos::base::v1beta1::Coin,
     >,
 }
-#[allow(clippy::derive_partial_eq_without_eq)]
-#[derive(Clone, PartialEq, ::prost::Message)]
-pub struct HandshakeMetadata {
-    #[prost(string, tag = "1")]
-    pub provider_fee_pool_addr: ::prost::alloc::string::String,
-    #[prost(string, tag = "2")]
-    pub version: ::prost::alloc::string::String,
-}
 /// SlashAcks contains cons addresses of consumer chain validators
-/// successfully slashed on the provider chain
+/// successfully slashed on the provider chain.
 #[allow(clippy::derive_partial_eq_without_eq)]
 #[derive(Clone, PartialEq, ::prost::Message)]
 pub struct SlashAcks {
     #[prost(string, repeated, tag = "1")]
     pub addresses: ::prost::alloc::vec::Vec<::prost::alloc::string::String>,
 }
-/// ConsumerAdditionProposals holds pending governance proposals on the provider chain to spawn a new chain.
+/// ConsumerAdditionProposals holds pending governance proposals on the provider
+/// chain to spawn a new chain.
 #[allow(clippy::derive_partial_eq_without_eq)]
 #[derive(Clone, PartialEq, ::prost::Message)]
 pub struct ConsumerAdditionProposals {
@@ -830,7 +740,8 @@ pub struct ConsumerAdditionProposals {
     #[prost(message, repeated, tag = "1")]
     pub pending: ::prost::alloc::vec::Vec<ConsumerAdditionProposal>,
 }
-/// ConsumerRemovalProposals holds pending governance proposals on the provider chain to remove (and stop) a consumer chain.
+/// ConsumerRemovalProposals holds pending governance proposals on the provider
+/// chain to remove (and stop) a consumer chain.
 #[allow(clippy::derive_partial_eq_without_eq)]
 #[derive(Clone, PartialEq, ::prost::Message)]
 pub struct ConsumerRemovalProposals {
@@ -894,6 +805,30 @@ pub struct VscSendTimestamp {
         super::super::super::super::google::protobuf::Timestamp,
     >,
 }
+/// ValidatorSetChangePackets is a pb list of ccv.ValidatorSetChangePacketData.
+#[allow(clippy::derive_partial_eq_without_eq)]
+#[derive(Clone, PartialEq, ::prost::Message)]
+pub struct ValidatorSetChangePackets {
+    #[prost(message, repeated, tag = "1")]
+    pub list: ::prost::alloc::vec::Vec<super::super::v1::ValidatorSetChangePacketData>,
+}
+/// MaturedUnbondingOps defines a list of ids corresponding to ids of matured
+/// unbonding operations.
+#[allow(clippy::derive_partial_eq_without_eq)]
+#[derive(Clone, PartialEq, ::prost::Message)]
+pub struct MaturedUnbondingOps {
+    #[prost(uint64, repeated, tag = "1")]
+    pub ids: ::prost::alloc::vec::Vec<u64>,
+}
+/// ExportedVscSendTimestamps is VscSendTimestamp with chainID info for exporting to genesis
+#[allow(clippy::derive_partial_eq_without_eq)]
+#[derive(Clone, PartialEq, ::prost::Message)]
+pub struct ExportedVscSendTimestamp {
+    #[prost(string, tag = "1")]
+    pub chain_id: ::prost::alloc::string::String,
+    #[prost(message, repeated, tag = "2")]
+    pub vsc_send_timestamps: ::prost::alloc::vec::Vec<VscSendTimestamp>,
+}
 #[allow(clippy::derive_partial_eq_without_eq)]
 #[derive(Clone, PartialEq, ::prost::Message)]
 pub struct KeyAssignmentReplacement {
@@ -905,7 +840,8 @@ pub struct KeyAssignmentReplacement {
     pub power: i64,
 }
 /// Used to serialize the ValidatorConsumerPubKey index from key assignment
-/// ValidatorConsumerPubKey: (chainID, providerAddr consAddr) -> consumerKey tmprotocrypto.PublicKey
+/// ValidatorConsumerPubKey: (chainID, providerAddr consAddr) -> consumerKey
+/// tmprotocrypto.PublicKey
 #[allow(clippy::derive_partial_eq_without_eq)]
 #[derive(Clone, PartialEq, ::prost::Message)]
 pub struct ValidatorConsumerPubKey {
@@ -917,7 +853,8 @@ pub struct ValidatorConsumerPubKey {
     pub consumer_key: ::core::option::Option<::tendermint_proto::crypto::PublicKey>,
 }
 /// Used to serialize the ValidatorConsumerAddr index from key assignment
-/// ValidatorByConsumerAddr: (chainID, consumerAddr consAddr) -> providerAddr consAddr
+/// ValidatorByConsumerAddr: (chainID, consumerAddr consAddr) -> providerAddr
+/// consAddr
 #[allow(clippy::derive_partial_eq_without_eq)]
 #[derive(Clone, PartialEq, ::prost::Message)]
 pub struct ValidatorByConsumerAddr {
@@ -950,7 +887,7 @@ pub struct QueryConsumerGenesisRequest {
 #[derive(Clone, PartialEq, ::prost::Message)]
 pub struct QueryConsumerGenesisResponse {
     #[prost(message, optional, tag = "1")]
-    pub genesis_state: ::core::option::Option<super::super::consumer::v1::GenesisState>,
+    pub genesis_state: ::core::option::Option<super::super::v1::ConsumerGenesisState>,
 }
 #[allow(clippy::derive_partial_eq_without_eq)]
 #[derive(Clone, PartialEq, ::prost::Message)]
@@ -1030,11 +967,12 @@ pub struct QueryThrottleStateResponse {
     /// current slash_meter state
     #[prost(int64, tag = "1")]
     pub slash_meter: i64,
-    /// allowance of voting power units (int) that the slash meter is given per replenish period
-    /// this also serves as the max value for the meter.
+    /// allowance of voting power units (int) that the slash meter is given per
+    /// replenish period this also serves as the max value for the meter.
     #[prost(int64, tag = "2")]
     pub slash_meter_allowance: i64,
-    /// next time the slash meter could potentially be replenished, iff it's not full
+    /// next time the slash meter could potentially be replenished, iff it's not
+    /// full
     #[prost(message, optional, tag = "3")]
     pub next_replenish_candidate: ::core::option::Option<
         super::super::super::super::google::protobuf::Timestamp,
@@ -1059,7 +997,8 @@ pub struct QueryThrottledConsumerPacketDataResponse {
     #[prost(message, repeated, tag = "3")]
     pub packet_data_instances: ::prost::alloc::vec::Vec<ThrottledPacketDataWrapper>,
 }
-/// A query wrapper type for the global entry and data relevant to a throttled slash packet.
+/// A query wrapper type for the global entry and data relevant to a throttled
+/// slash packet.
 #[allow(clippy::derive_partial_eq_without_eq)]
 #[derive(Clone, PartialEq, ::prost::Message)]
 pub struct ThrottledSlashPacket {
@@ -1068,7 +1007,8 @@ pub struct ThrottledSlashPacket {
     #[prost(message, optional, tag = "2")]
     pub data: ::core::option::Option<super::super::v1::SlashPacketData>,
 }
-/// ThrottledPacketDataWrapper contains either SlashPacketData or VSCMaturedPacketData
+/// ThrottledPacketDataWrapper contains either SlashPacketData or
+/// VSCMaturedPacketData
 #[allow(clippy::derive_partial_eq_without_eq)]
 #[derive(Clone, PartialEq, ::prost::Message)]
 pub struct ThrottledPacketDataWrapper {
@@ -1375,7 +1315,8 @@ pub mod query_client {
                 );
             self.inner.unary(req, path, codec).await
         }
-        /// QueryThrottleState returns the main on-chain state relevant to currently throttled slash packets
+        /// QueryThrottleState returns the main on-chain state relevant to currently
+        /// throttled slash packets
         pub async fn query_throttle_state(
             &mut self,
             request: impl tonic::IntoRequest<super::QueryThrottleStateRequest>,
@@ -1406,8 +1347,8 @@ pub mod query_client {
                 );
             self.inner.unary(req, path, codec).await
         }
-        /// QueryThrottledConsumerPacketData returns a list of pending packet data instances
-        /// (slash packet and vsc matured) for a single consumer chain
+        /// QueryThrottledConsumerPacketData returns a list of pending packet data
+        /// instances (slash packet and vsc matured) for a single consumer chain
         pub async fn query_throttled_consumer_packet_data(
             &mut self,
             request: impl tonic::IntoRequest<
@@ -1440,7 +1381,8 @@ pub mod query_client {
                 );
             self.inner.unary(req, path, codec).await
         }
-        /// QueryRegisteredConsumerRewardDenoms returns a list of consumer reward denoms that are registered
+        /// QueryRegisteredConsumerRewardDenoms returns a list of consumer reward
+        /// denoms that are registered
         pub async fn query_registered_consumer_reward_denoms(
             &mut self,
             request: impl tonic::IntoRequest<
@@ -1535,7 +1477,8 @@ pub mod query_server {
             tonic::Response<super::QueryValidatorProviderAddrResponse>,
             tonic::Status,
         >;
-        /// QueryThrottleState returns the main on-chain state relevant to currently throttled slash packets
+        /// QueryThrottleState returns the main on-chain state relevant to currently
+        /// throttled slash packets
         async fn query_throttle_state(
             &self,
             request: tonic::Request<super::QueryThrottleStateRequest>,
@@ -1543,8 +1486,8 @@ pub mod query_server {
             tonic::Response<super::QueryThrottleStateResponse>,
             tonic::Status,
         >;
-        /// QueryThrottledConsumerPacketData returns a list of pending packet data instances
-        /// (slash packet and vsc matured) for a single consumer chain
+        /// QueryThrottledConsumerPacketData returns a list of pending packet data
+        /// instances (slash packet and vsc matured) for a single consumer chain
         async fn query_throttled_consumer_packet_data(
             &self,
             request: tonic::Request<super::QueryThrottledConsumerPacketDataRequest>,
@@ -1552,7 +1495,8 @@ pub mod query_server {
             tonic::Response<super::QueryThrottledConsumerPacketDataResponse>,
             tonic::Status,
         >;
-        /// QueryRegisteredConsumerRewardDenoms returns a list of consumer reward denoms that are registered
+        /// QueryRegisteredConsumerRewardDenoms returns a list of consumer reward
+        /// denoms that are registered
         async fn query_registered_consumer_reward_denoms(
             &self,
             request: tonic::Request<super::QueryRegisteredConsumerRewardDenomsRequest>,
@@ -2130,9 +2074,7 @@ pub struct GenesisState {
     pub unbonding_ops: ::prost::alloc::vec::Vec<UnbondingOp>,
     /// empty for a new chain
     #[prost(message, optional, tag = "4")]
-    pub mature_unbonding_ops: ::core::option::Option<
-        super::super::v1::MaturedUnbondingOps,
-    >,
+    pub mature_unbonding_ops: ::core::option::Option<MaturedUnbondingOps>,
     /// empty for a new chain
     #[prost(message, repeated, tag = "5")]
     pub valset_update_id_to_height: ::prost::alloc::vec::Vec<ValsetUpdateIdToHeight>,
@@ -2153,8 +2095,14 @@ pub struct GenesisState {
     /// empty for a new chain
     #[prost(message, repeated, tag = "11")]
     pub consumer_addrs_to_prune: ::prost::alloc::vec::Vec<ConsumerAddrsToPrune>,
+    #[prost(message, repeated, tag = "12")]
+    pub init_timeout_timestamps: ::prost::alloc::vec::Vec<InitTimeoutTimestamp>,
+    #[prost(message, repeated, tag = "13")]
+    pub exported_vsc_send_timestamps: ::prost::alloc::vec::Vec<ExportedVscSendTimestamp>,
 }
-/// consumer chain
+/// The provider CCV module's knowledge of consumer state.
+///
+/// Note this type is only used internally to the provider CCV module.
 #[allow(clippy::derive_partial_eq_without_eq)]
 #[derive(Clone, PartialEq, ::prost::Message)]
 pub struct ConsumerState {
@@ -2172,17 +2120,17 @@ pub struct ConsumerState {
     pub initial_height: u64,
     /// ConsumerGenesis defines the initial consumer chain genesis states
     #[prost(message, optional, tag = "5")]
-    pub consumer_genesis: ::core::option::Option<
-        super::super::consumer::v1::GenesisState,
-    >,
-    /// PendingValsetChanges defines the pending validator set changes for the consumer chain
+    pub consumer_genesis: ::core::option::Option<super::super::v1::ConsumerGenesisState>,
+    /// PendingValsetChanges defines the pending validator set changes for the
+    /// consumer chain
     #[prost(message, repeated, tag = "6")]
     pub pending_valset_changes: ::prost::alloc::vec::Vec<
         super::super::v1::ValidatorSetChangePacketData,
     >,
     #[prost(string, repeated, tag = "7")]
     pub slash_downtime_ack: ::prost::alloc::vec::Vec<::prost::alloc::string::String>,
-    /// UnbondingOpsIndex defines the unbonding operations waiting on this consumer chain
+    /// UnbondingOpsIndex defines the unbonding operations waiting on this consumer
+    /// chain
     #[prost(message, repeated, tag = "8")]
     pub unbonding_ops_index: ::prost::alloc::vec::Vec<VscUnbondingOps>,
 }
